Changed all Longword defs to UInt32 since under iOS64 a Longword is a UInt64.

+++ source/hpp/iOSDevice64/Release/AbDfStrm.hpp
@@ -121,13 +121,13 @@ private:
 	Abdfbase::TAbLogger* FLog;
 	Abdfinw::TAbDfInputWindow* FSlideWin;
 	__int64 FStartOfs;
-	unsigned long FStoredSize;
+	unsigned FStoredSize;
 	System::Byte *FStream;
 	System::Byte *FStrmEnd;
 	bool FUseDeflate64;
 	
 protected:
-	unsigned long __fastcall lzsGetApproxSize(void);
+	unsigned __fastcall lzsGetApproxSize(void);
 	int __fastcall lzsGetStaticSize(void);
 	int __fastcall lzsGetStoredSize(void);
 	bool __fastcall lzsIsFull(void);
